+++ src/lib.rs
@@ -1,27 +1,13 @@
 #![allow(clippy::needless_return)]
+//! Crate for reading and manipulating `NITF` files
 
-//! Crate for reading and manipulating NITF files
-use std::fs::File;
-use std::path::Path;
+/// General `NIFT` interface functions and types
+pub mod nitf;
 
-pub mod nitf_2_1;
+/// Functions and types specific to `SICD` files
+#[cfg(feature = "sicd")]
 pub mod sicd;
 
-use nitf_2_1::Nitf;
-
-/// Construct a `Nitf` object from a file `path`.
-///
-/// # Example
-///
-///     use std::path::Path;
-///     use nitf_rs::read_nitf;
-///
-///     let nitf_path = Path::new("../example.nitf");
-///     let nitf = read_nitf(nitf_path);
-pub fn read_nitf(path: &Path) -> Nitf {
-    let mut reader = File::open(path).unwrap();
-    return Nitf::from_file(&mut reader);
-}
 
 // UNIT TESTS
 #[cfg(test)]