+++ src/sicd/mod.rs
@@ -1,61 +1,43 @@
 //! Sensor Independent Complex Data support
+use std::fs::File;
+use std::path::Path;
+use quick_xml::de::from_str;
 
-use serde::Deserialize;
+use crate::nitf::Nitf;
+
+use self::sicd_types::SicdMeta;
 
 // TODO: Write tests
 pub mod sicd_types;
-use sicd_types::{
-    antenna::Antenna,
-    collection_info::CollectionInfo,
-    error_statistics::ErrorStatistics,
-    geo_data::GeoData,
-    grid::Grid,
-    image_creation::ImageCreation,
-    image_data::ImageData,
-    image_formation::{ImageFormation, RgAzComp, Rma},
-    match_info::MatchInfo,
-    pfa::Pfa,
-    position::Position,
-    radar_collection::RadarCollection,
-    radiometric::Radiometric,
-    scpcoa::ScpCoa,
-    timeline::Timeline,
-};
 
-#[derive(Debug, Deserialize, PartialEq, Clone)]
+/// SICD file structure
+#[derive(Debug)]
 pub struct Sicd {
-    #[serde(rename = "CollectionInfo")]
-    pub collection_info: CollectionInfo,
-    #[serde(rename = "ImageCreation")]
-    pub image_creation: Option<ImageCreation>,
-    #[serde(rename = "ImageData")]
-    pub image_data: ImageData,
-    #[serde(rename = "GeoData")]
-    pub geo_data: GeoData,
-    #[serde(rename = "Grid")]
-    pub grid: Grid,
-    #[serde(rename = "Timeline")]
-    pub timeline: Timeline,
-    #[serde(rename = "Position")]
-    pub position: Position,
-    #[serde(rename = "RadarCollection")]
-    pub radar_collection: RadarCollection,
-    #[serde(rename = "ImageFormation")]
-    pub image_formation: ImageFormation,
-    #[serde(rename = "SCPCOA")]
-    pub scpcoa: ScpCoa,
-    #[serde(rename = "Radiometric")]
-    pub radiometric: Option<Radiometric>,
-    #[serde(rename = "Antenna")]
-    pub antenna: Option<Antenna>,
-    #[serde(rename = "ErrorStatistics")]
-    pub error_statistics: Option<ErrorStatistics>,
-    #[serde(rename = "MatchInfo")]
-    pub match_info: Option<MatchInfo>,
-    #[serde(rename = "RgAzComp")]
-    pub rg_az_comp: Option<RgAzComp>,
-    #[serde(rename = "PFA")]
-    pub pfa: Option<Pfa>,
-    #[serde(rename = "RMA")]
-    pub rma: Option<Rma>,
+    /// Nitf file object and associated metadata
+    pub nitf: Nitf,
+    /// Parsed SICD xml metadata
+    pub sicd_meta: SicdMeta,
+}
+
+/// Construct a [Sicd] object from a file `path`.
+///
+/// # Example
+///
+///     use std::path::Path;
+///     use nitf_rs::read_sicd;
+///
+///     let sicd_path = Path::new("../example.nitf");
+///     let sicd = read_sicd(sicd_path);
+pub fn read_sicd(path: &Path) -> Sicd {
+    let mut reader = File::open(path).unwrap();
+    return Sicd::from_file(&mut reader);
+}
+
+impl Sicd {
+    pub fn from_file(reader: &mut File) -> Self {
+        let nitf = Nitf::from_file(reader);
+        let sicd_str = String::from_utf8(nitf.data_extension_segments[0].data[..].to_vec()).unwrap();
+        let sicd_meta: SicdMeta = from_str(&sicd_str).unwrap();
+        Self { nitf, sicd_meta }
+    }
 }