+++ src/sicd/sicd_types/mod.rs
@@ -18,6 +18,61 @@ pub mod radiometric;
 pub mod scpcoa;
 pub mod timeline;
 
+use antenna::Antenna;
+use collection_info::CollectionInfo;
+use error_statistics::ErrorStatistics;
+use geo_data::GeoData;
+use grid::Grid;
+use image_creation::ImageCreation;
+use image_data::ImageData;
+use image_formation::{ImageFormation, RgAzComp, Rma};
+use match_info::MatchInfo;
+use pfa::Pfa;
+use position::Position;
+use radar_collection::RadarCollection;
+use radiometric::Radiometric;
+use scpcoa::ScpCoa;
+use timeline::Timeline;
+
+
+#[derive(Debug, Deserialize, PartialEq, Clone)]
+pub struct SicdMeta {
+    #[serde(rename = "CollectionInfo")]
+    pub collection_info: CollectionInfo,
+    #[serde(rename = "ImageCreation")]
+    pub image_creation: Option<ImageCreation>,
+    #[serde(rename = "ImageData")]
+    pub image_data: ImageData,
+    #[serde(rename = "GeoData")]
+    pub geo_data: GeoData,
+    #[serde(rename = "Grid")]
+    pub grid: Grid,
+    #[serde(rename = "Timeline")]
+    pub timeline: Timeline,
+    #[serde(rename = "Position")]
+    pub position: Position,
+    #[serde(rename = "RadarCollection")]
+    pub radar_collection: RadarCollection,
+    #[serde(rename = "ImageFormation")]
+    pub image_formation: ImageFormation,
+    #[serde(rename = "SCPCOA")]
+    pub scpcoa: ScpCoa,
+    #[serde(rename = "Radiometric")]
+    pub radiometric: Option<Radiometric>,
+    #[serde(rename = "Antenna")]
+    pub antenna: Option<Antenna>,
+    #[serde(rename = "ErrorStatistics")]
+    pub error_statistics: Option<ErrorStatistics>,
+    #[serde(rename = "MatchInfo")]
+    pub match_info: Option<MatchInfo>,
+    #[serde(rename = "RgAzComp")]
+    pub rg_az_comp: Option<RgAzComp>,
+    #[serde(rename = "PFA")]
+    pub pfa: Option<Pfa>,
+    #[serde(rename = "RMA")]
+    pub rma: Option<Rma>,
+}
+
 #[derive(Debug, Deserialize, PartialEq, Clone)]
 pub struct RowCol {
     #[serde(rename = "Row")]